lang: Allow matching underscores in some of the identifier's
This allows matching underscores in some of the identifier's, but not
when they're the last character.

This caused me to suffer a bit of pain tracking down a bug which turned
out to be in the lexer. It started with a failing test that I wrote in:

https://github.com/blynn/nex/commit/974c2498c46eaf4fb3963cfc059bcefd66ea48a8

and which followed with a fix in:

https://github.com/blynn/nex/commit/52682f463a45fdc37630e650d82f126a32d810a3

Glad that's fixed!